fix(devices): Don't attempt to delete config for Huron devices
Huron devices don't store their config with Lyra, so we shouldn't attempt to delete it when deleting
a Huron device

Discovered during investigation of SPARK-26055

+++ app/modules/squared/devices/remDevice/remDevice.js
@@ -34,7 +34,7 @@
         rdc.deleteDevice = function () {
           var deleteConfigPromise;
 
-          if (rdc.deleteConfig || !rdc.deleteConfigurationOptionAvailable) {
+          if (rdc.device.isCloudberryDevice() && (rdc.deleteConfig || !rdc.deleteConfigurationOptionAvailable)) {
             deleteConfigPromise = CsdmLyraConfigurationService.deleteConfig(rdc.device.cisUuid, rdc.device.wdmUrl);
           } else {
             deleteConfigPromise = $q.resolve();
